Move waste-balances repository into module
