+++ src/test/create-test-server.js
@@ -31,7 +31,7 @@ import { createInMemoryFormSubmissionsRepositoryPlugin } from '#repositories/for
 import { createInMemoryOrganisationsRepositoryPlugin } from '#repositories/organisations/inmemory.plugin.js'
 import { createInMemorySummaryLogsRepositoryPlugin } from '#repositories/summary-logs/inmemory.plugin.js'
 import { createInMemorySystemLogsRepositoryPlugin } from '#repositories/system-logs/inmemory.plugin.js'
-import { createInMemoryWasteBalancesRepositoryPlugin } from '#repositories/waste-balances/inmemory.plugin.js'
+import { createInMemoryWasteBalancesRepositoryPlugin } from '#waste-balances/repository/inmemory.plugin.js'
 import { createInMemoryWasteRecordsRepositoryPlugin } from '#repositories/waste-records/inmemory.plugin.js'
 
 /**
