+++ src/waste-balances/index.js
@@ -1,0 +1,12 @@
+/**
+ * Waste Balances Module
+ *
+ * This module contains the repository layer for waste balances.
+ * Domain and routes remain in their original locations and will move in
+ * subsequent slices.
+ *
+ * @module waste-balances
+ */
+
+// Repository exports
+export { createWasteBalancesRepository } from './repository/mongodb.js'
